Add docker Change getPriceForShop and getPriceForVendor param
