+++ .gitignore
@@ -271,6 +271,7 @@ cython_debug/
 /.idea/encodings.xml
 /.idea/jarRepositories.xml
 /src/server/java/data-com.gr3530904_90104.service.iml
+/src/server/data-service/.gradle/
 
 # Headers and cookies config
 /src/server/aggregator/config.py
