+++ src/server/aggregator/Dockerfile
@@ -1,0 +1,30 @@
+FROM python:3.9
+
+# install google chrome
+RUN wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add -
+RUN sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
+RUN apt-get -y update
+RUN apt-get install -y google-chrome-stable
+
+# install chromedriver
+RUN apt-get install -yqq unzip
+RUN wget -O /tmp/chromedriver.zip http://chromedriver.storage.googleapis.com/`curl -sS chromedriver.storage.googleapis.com/LATEST_RELEASE`/chromedriver_linux64.zip
+RUN unzip /tmp/chromedriver.zip chromedriver -d /usr/local/bin/
+
+# set display port to avoid crash
+ENV DISPLAY=:99
+
+RUN apt-get update
+RUN apt-get -y install cron
+RUN apt-get -y install pip
+
+WORKDIR /aggregator
+COPY data-aggregator.py .
+COPY config.py .
+COPY requirements.txt .
+
+RUN pip install -r requirements.txt
+
+#RUN crontab -l | { cat; echo "* * * * * python3.9 data-aggregator.py"; } | crontab -
+#CMD cron
+CMD ["python", "data-aggregator.py"]
