+++ src/server/aggregator/requirements.txt
@@ -1,0 +1,28 @@
+async-generator==1.10
+attrs==22.1.0
+beautifulsoup4==4.11.1
+certifi==2022.12.7
+cffi==1.15.1
+charset-normalizer==2.1.1
+colorama==0.4.6
+exceptiongroup==1.0.4
+h11==0.14.0
+idna==3.4
+outcome==1.2.0
+packaging==22.0
+pycparser==2.21
+PySocks==1.7.1
+python-dotenv==0.21.0
+requests==2.28.1
+selenium==4.7.2
+selenium-stealth==1.0.6
+sniffio==1.3.0
+sortedcontainers==2.4.0
+soupsieve==2.3.2.post1
+tqdm==4.64.1
+trio==0.22.0
+trio-websocket==0.9.2
+urllib3==1.26.12
+webdriver-manager==3.8.5
+wsproto==1.2.0
+schedule==1.1.0
