+++ src/server/data-service/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sun Dec 25 17:08:50 MSK 2022
+gradle.version=7.5.1
