+++ src/server/data-service/Dockerfile
@@ -1,0 +1,5 @@
+FROM gradle:jdk17
+WORKDIR /data-service
+COPY build.gradle .
+COPY ./src ./src
+ENTRYPOINT ["gradle", "run"]
