+++ src/server/data-service/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'data-service'
+
