+++ src/server/data-service/src/main/java/com/gr3530904_90104/table/Architecture.java
