+++ src/server/data-service/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+spring.datasource.url=jdbc:postgresql://173.18.0.2:5432/postgres
+spring.datasource.username=video_cards_data_user
+spring.datasource.password=video_cards_data_password
+spring.datasource.driver-class-name=org.postgresql.Driver
+
+spring.jpa.show-sql=true
+
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.format_sql=true
+server.port=8080
+DEBUG=true
+logging.level.org.springframework.web=DEBUG
