--- src/server/java/resouces/application.properties
@@ -1,11 +1,0 @@
-DEBUG=true
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
-spring.datasource.username=postgres
-spring.datasource.password=12345678
-spring.datasource.driver-class-name=org.postgresql.Driver
-
-spring.jpa.show-sql=true
-
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.properties.hibernate.format_sql=true
